pyproject.toml
[tool.poetry]
name = "ai-assistant-interlink"
version = "0.1.0"
description = ""
authors = ["Joel <joelkronander@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.28.1"
openai = "^1.1.2"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
